// src/constants/enums/bin-setting.ts
export enum BinColor {
  Yellow = '#FFE97A',
  Red = '#E01A22',
  LightGreen = '#60D260',
  DarkGreen = '#1D6339',
  Purple = '#7D26C5',
  Maroon = '#9C1235',
  Blue = '#1867C1',
  BatteryBack = '#E8E8E8',
  Pink = '#FFE6E4',
}

export enum BinKey {
  Yellow = 'yellow',
  Red = 'red',
  LightGreen = 'lightGreen',
  DarkGreen = 'darkGreen',
  Purple = 'purple',
  Maroon = 'maroon',
  Blue = 'blue',
  BatteryBack = 'batteryBack',
  RedCycle = 'redCycle',
}

export enum BinRecycleType {
  Glass = 'recycleGlass',
  Mixed = 'recycleMixed',
  Paper = 'recyclePaper',
  Plastic = 'recyclePlastic',
  Waste = 'recycleWaste',
  Other = 'recycleOther',
  RedCycle = 'recycleRedCycle',
}